tests/wpt/web-platform-tests/content-security-policy/worker-src/dedicated-none.sub.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../support/testharness-helper.js"></script>
<meta http-equiv="Content-Security-Policy" content="worker-src 'none'">
<script>
  var url = new URL("../support/ping.js", document.baseURI).toString();
  assert_worker_is_blocked(url, "Same-origin dedicated worker blocked by host-source expression.");

  var b = new Blob(["postMessage('ping');"], {type: "text/javascript"});
  var url = URL.createObjectURL(b);
  assert_worker_is_blocked(url, "blob: dedicated worker blocked by 'blob:'.");
</script>
