LayoutTests/imported/w3c/web-platform-tests/content-security-policy/worker-src/dedicated-none.sub-expected.txt


FAIL Same-origin dedicated worker blocked by host-source expression. assert_throws_dom: function "function () {
      var w = new Worker(url);
    }" did not throw
FAIL blob: dedicated worker blocked by 'blob:'. assert_throws_dom: function "function () {
      var w = new Worker(url);
    }" did not throw
